+++ examples/fingerprint.rs
@@ -1,4 +1,4 @@
-use failure::Error;
+use osshkeys::error::OsshResult;
 use osshkeys::keys::FingerprintHash;
 use osshkeys::{PublicKey, PublicParts};
 use std::env;
@@ -6,7 +6,7 @@ use std::fmt::Display;
 use std::fs;
 use std::path::Path;
 
-fn print_fingerprint<P: Display + AsRef<Path>>(path: P) -> Result<(), Error> {
+fn print_fingerprint<P: Display + AsRef<Path>>(path: P) -> OsshResult<()> {
     print!("{}: ", path);
     match fs::read_to_string(path) {
         Ok(s) => {
@@ -23,7 +23,7 @@ fn print_fingerprint<P: Display + AsRef<Path>>(path: P) -> Result<(), Error> {
     Ok(())
 }
 
-fn main() -> Result<(), Error> {
+fn main() -> OsshResult<()> {
     let mut argv = env::args();
     argv.next();
     for arg in argv {